// frontend/src/types/Books.ts
export interface Book {
    bookID: number;         // matches "bookID"
    title: string;          // matches "title"
    author: string;         // matches "author"
    publisher: string;      // matches "publisher"
    isbn: string;           // matches "isbn"
    classification: string; // matches "classification"
    category: string;       // matches "category"
    pageCount: number;      // matches "pageCount"
    price: number;          // matches "price"
  }
  